fix(P0): FX-13 migrar Double a BigDecimal para precision monetaria
FX-13 (FUNCIONAL/CRITICO): TarjetaCombustible.saldo y Recorrido.importeAbastecido
usaban Double para aritmetica monetaria, lo que produce perdida de precision
floating-point (ej. 0.1 + 0.2 != 0.3).

Migracion a BigDecimal (precision=12, scale=2):

Entidades:
- TarjetaCombustible.saldo: Double -> BigDecimal
- Recorrido.importeAbastecido: Double -> BigDecimal

DTOs:
- RecorridoRequest.importeAbastecido: Double -> BigDecimal
- RecorridoResponse.importeAbastecido: Double -> BigDecimal
- TarjetaCombustibleRequest.saldo: Double -> BigDecimal
- TarjetaCombustibleResponse.saldo: Double -> BigDecimal

Repository:
- RecorridoRepository.sumCostoCombustible: Double -> BigDecimal

Service:
- RecorridoServiceImpl.create/update/delete: aritmetica con BigDecimal
  (compareTo, subtract, add) en lugar de operadores (+, -, <=).
- ReporteDashboardEjecutivoServiceImpl: elimina conversion intermedia
  BigDecimal.valueOf(Double) al recibir BigDecimal directamente.

Flyway:
- Agregada dependencia flyway-core + flyway-database-postgresql.
- Migracion V1__fx13_bigdecimal_and_shedlock.sql convierte columnas
  DOUBLE a DECIMAL(12,2) con CAST y crea tabla shedlock (FX-29).
- Configuracion spring.flyway.* en application.properties.
- baseline-on-migrate=true para BDs existentes.

Refs: plan-defectos.md seccion 3 (P0) FX-13

+++ fleet-management/src/main/java/com/fleet/management/dto/tarjetacombustible/TarjetaCombustibleRequest.java
@@ -6,6 +6,8 @@ import jakarta.validation.constraints.Positive;
 import jakarta.validation.constraints.Size;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -19,7 +21,8 @@ public class TarjetaCombustibleRequest {
 
     @NotNull(message = "El saldo es obligatorio")
     @Positive(message = "El saldo debe ser mayor a cero")
-    private Double saldo;
+    // FX-13: migrado de Double a BigDecimal para precision monetaria.
+    private BigDecimal saldo;
 
     @NotNull(message = "La moneda es obligatoria")
     private Long currencyId;
